src/back_end/requirements.txt
click==8.1.3
dotmap==1.3.30
enum-compat==0.0.3
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
meshtastic==2.0.3
pexpect==4.8.0
protobuf==4.21.9
ptyprocess==0.7.0
pygatt==4.0.5
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
PyYAML==6.0
tabulate==0.9.0
timeago==1.0.16
Werkzeug==2.2.2
zipp==3.10.0
